fix(extensions): moved yfm table to yfm
