--- src/extensions/behavior/YfmTableAdditions/YfmTableControls/index.ts
@@ -1,10 +1,0 @@
-import {Extension} from '../../../../core';
-
-import {tableControlsPlugin} from './buttons';
-
-export {tableControlsPlugin} from './buttons';
-
-// TODO: split: plus buttons extenion & cell controls extension
-export const YfmTableControlsExtension: Extension = (builder) => {
-    builder.addPlugin(tableControlsPlugin);
-};
