--- src/extensions/behavior/YfmTableAdditions/const.ts
@@ -1,1 +1,0 @@
-export {YfmTableNode} from '../../../extensions/yfm';
