--- src/extensions/behavior/YfmTableAdditions/index.ts
@@ -1,7 +1,0 @@
-import {ExtensionAuto} from '../../../core';
-
-import {tableControlsPlugin} from './YfmTableControls';
-
-export const YfmTableAdditions: ExtensionAuto = (builder) => {
-    builder.addPlugin(tableControlsPlugin);
-};
