+++ src/extensions/yfm/YfmTable/index.ts
@@ -1,5 +1,3 @@
-import {Plugin} from 'prosemirror-state';
-
 import type {Action, ExtensionWithOptions} from '../../../core';
 import {goToNextCell} from '../../../table-utils';
 
@@ -7,7 +5,8 @@ import {YfmTableSpecs, YfmTableSpecsOptions} from './YfmTableSpecs';
 import {createYfmTable} from './actions';
 import {backspaceCommand} from './commands/backspace';
 import {goToNextRow} from './commands/goToNextRow';
-import {fixPastedTableBodies} from './paste';
+import {tableControlsPlugin} from './plugins/YfmTableControls/buttons';
+import {tableTransformPastedPlugin} from './plugins/tableTransformPastedPlugin';
 
 const action = 'createYfmTable';
 
@@ -25,27 +24,16 @@ export type YfmTableOptions = YfmTableSpecsOptions & {};
 export const YfmTable: ExtensionWithOptions<YfmTableOptions> = (builder, options) => {
     builder.use(YfmTableSpecs, options);
 
-    builder
-        .addKeymap(() => ({
-            Tab: goToNextCell('next'),
-            'Shift-Tab': goToNextCell('prev'),
-            ArrowDown: goToNextRow('down'),
-            ArrowUp: goToNextRow('up'),
-            Backspace: backspaceCommand,
-        }))
-
-        .addAction(action, () => createYfmTable);
-
-    builder.addPlugin(
-        ({schema}) =>
-            new Plugin({
-                props: {
-                    transformPasted(slice) {
-                        return fixPastedTableBodies(slice, schema);
-                    },
-                },
-            }),
-    );
+    builder.addKeymap(() => ({
+        Tab: goToNextCell('next'),
+        'Shift-Tab': goToNextCell('prev'),
+        ArrowDown: goToNextRow('down'),
+        ArrowUp: goToNextRow('up'),
+        Backspace: backspaceCommand,
+    }));
+    builder.addAction(action, () => createYfmTable);
+    builder.addPlugin(tableTransformPastedPlugin);
+    builder.addPlugin(tableControlsPlugin);
 };
 
 declare global {
