+++ src/extensions/yfm/YfmTable/plugins/YfmTableControls/view.scss
