+++ src/extensions/yfm/YfmTable/plugins/tableTransformPastedPlugin.ts
@@ -1,0 +1,16 @@
+import {Schema} from 'prosemirror-model';
+import {Plugin} from 'prosemirror-state';
+
+import {fixPastedTableBodies} from '../paste';
+
+interface Args {
+    schema: Schema;
+}
+export const tableTransformPastedPlugin = ({schema}: Args) =>
+    new Plugin({
+        props: {
+            transformPasted(slice) {
+                return fixPastedTableBodies(slice, schema);
+            },
+        },
+    });
